requirements.txt
cffi==0.8.6
cryptography==0.6.1
pyOpenSSL==0.14
pycparser==2.10
pycrypto==2.6.1
six==1.8.0
wsgiref==0.1.2
